graph working, styling still to do
