+++ app/javascript/packs/application.js
@@ -3,6 +3,8 @@
 // a relevant structure within app/javascript and only use these pack files to reference
 // that code so it'll be compiled.
 //= require algolia/v3/algoliasearch.min
+//= require chartkick
+//= require Chart.bundle
 
 require("@rails/ujs").start()
 require("turbolinks").start()
