+++ app/javascript/packs/chartkick.js
@@ -1,0 +1,2 @@
+require("chartkick")
+require("chart.js")
