+++ package.json
@@ -8,6 +8,8 @@
     "@rails/webpacker": "4.3.0",
     "algoliasearch": "^4.8.5",
     "bootstrap": "^4.6.0",
+    "chart.js": "^2.9.4",
+    "chartkick": "^3.2.1",
     "jquery": "^3.5.1",
     "popper.js": "^1.16.1",
     "turbolinks": "^5.2.0"
